chore: introduce corba as the cli scafford (#773)
related #739
