+++ api/test/shell/cli_test.sh
@@ -40,7 +40,7 @@ clean_logfile() {
 trap clean_up EXIT
 
 export GO111MODULE=on
-go build -o ./manager-api -ldflags "-X main.Version=${VERSION}" .
+go build -o ./manager-api -ldflags "-X github.com/apisix/manager-api/cmd.Version=${VERSION}" .
 
 # default level: warn, path: logs/error.log
 
@@ -155,7 +155,7 @@ sleep 6
 
 cat ${logfile}
 
-if [[ `grep -c "api/main.go:" ${logfile}` -ne '1' ]]; then
+if [[ `grep -c "cmd/managerapi.go" ${logfile}` -ne '1' ]]; then
     echo "failed: failed to write the correct caller"
     exit 1
 fi
